+++ UserApplication/Enemy/EnemyHP2DUI.cpp
@@ -34,9 +34,9 @@ void EnemyHP2DUI::Update()
 	HPBackSprite->SetSize(BackHpDownSize);
 }
 
-void EnemyHP2DUI::EnemyHpUpdate(const uint32_t& nowHp,const uint32_t& MaxHp)
+void EnemyHP2DUI::EnemyHpUpdate(const float& nowHp,const float& MaxHp)
 {
-	HpSize.x = HpBarMaxSize * ( static_cast< float >( nowHp ) / static_cast< float >( MaxHp ) );
+	HpSize.x = HpBarMaxSize * ( nowHp  / MaxHp );
 	HP->SetSize(HpSize);
 	hpUpdate->EasingMaterial(HpSize.x);
 }
